requirements.txt
fastapi==0.115.2
uvicorn==0.23.2
httpx==0.27.2
pydantic==2.3.0
websockets==12.0

torch==2.2.0
torchaudio==2.2.0
numpy==1.26.4

python-dotenv==1.0.1
tqdm==4.66.0

chromadb==0.4.22
sentence-transformers==2.3.1
transformers==4.36.0
accelerate>=0.20.0

# Additional dependencies
aiohttp>=3.9.0
aiofiles>=23.0.0

# Optional - remove if not needed for now
# whisper==2023.12.15
# sounddevice==0.8.4
# pyttsx3==2.90
# playsound==1.3.0
# TTS==0.13.1
# redis==5.0.10